Address CodeRabbit review comments for SNI crash fix
Improvements to route clearing logic:
- Reset to default domain after clearing SNI routes in uws_app_clear_sni_routes()
- Allow clearSNIRoutes() to be called even with empty server name list
- Preallocate ArrayList capacity in clearSNIRoutesHelper()
- Add error logging when server name append fails

Test improvements:
- Rename test file from 21792-sni-crash.test.ts to 21792.test.ts
- Add tls.servername to all fetch calls for proper SNI testing
- Change from hardcoded port: 18443 to dynamic port: 0
- Un-skip all tests (all 3 now pass)
